Move colorPalette to shared/constants
