+++ app/scenes/Document/components/RevisionViewer.tsx
@@ -1,6 +1,6 @@
 import { observer } from "mobx-react";
 import * as React from "react";
-import { colorPalette } from "@shared/utils/collections";
+import { colorPalette } from "@shared/constants";
 import type Document from "~/models/Document";
 import type Revision from "~/models/Revision";
 import type { Props as EditorProps } from "~/components/Editor";
