+++ shared/constants.ts
@@ -10,6 +10,21 @@ import {
 
 export const MAX_AVATAR_DISPLAY = 6;
 
+/** Preset colors offered when choosing an icon color. */
+export const colorPalette = [
+  "#4E5C6E",
+  "#0366D6",
+  "#2BC2FF",
+  "#9E5CF7",
+  "#FF825C",
+  "#FF5C80",
+  "#FFBE0B",
+  "#42DED1",
+  "#00D084",
+  "#FF4DFA",
+  "#2F362F",
+];
+
 export const Pagination = {
   defaultLimit: 25,
   defaultOffset: 0,
